refactor(planner): Remove redundant files
