--- modules/planner/hooks/planManipulatorUtils.ts
@@ -1,21 +1,0 @@
-/**
- * Move the item at the given start index to the given end index.
- *
- * @param courses The semester to reorder
- * @param startIndex The starting index of the item to move
- * @param endIndex The destination index of the item to move
- */
-export function reorderSemester(courses: string[], startIndex: number, endIndex: number): string[] {
-  const result = Array.from(courses);
-  const [removed] = result.splice(startIndex, 1);
-  result.splice(endIndex, 0, removed);
-  return result;
-}
-
-export function reorderList<T>(list: T[], startIndex: number, endIndex: number) {
-  const result = Array.from(list);
-  const [removed] = result.splice(startIndex, 1);
-  result.splice(endIndex, 0, removed);
-
-  return result;
-}
